pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "calendar-sync"
version = "0.1.0"
description = "Synchronize calendars between Exchange EWS and Microsoft 365"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "msal>=1.34.0",
    "msal-extensions>=1.2.0",
    "Office365-REST-Python-Client>=2.6.0",
    "exchangelib>=5.6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pytz>=2024.1",
    "selenium>=4.0.0",
    "webdriver-manager>=4.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
calendar-sync = "calendar_sync.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
